Completion and BagFix Users
تکمیل بهش کاربران و فیکس باگ آپدیت ، حذف و دریافت اطلاعات کاربران

+++ server/src/users/users.service.ts
@@ -5,22 +5,26 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Users } from 'src/schemas/users.schema';
 import { hash } from 'bcrypt';
+import { Request } from 'express';
 
 @Injectable()
 export class UsersService {
   constructor(@InjectModel(Users.name) private UsersModel: Model<Users>) {}
 
   // ایجاد کاربر جدید
-  async create(createUserDto: CreateUserDto, req?: any) {
+  async create(createUserDto: CreateUserDto) {
     const { username, pass, email } = createUserDto;
 
+    // دریافت کاربر از نام کاربری و ایمیل
     const findUsername = await this.findByUsername(username);
     const findEmail = await this.findByEmail(email);
 
+    // شرط وجود کاربر
     if (findUsername.statusCode == 200 || findEmail.statusCode == 200) {
       return false;
     }
 
+    // هش پسورد
     const hashPass = await hash(pass, 10);
 
     await this.UsersModel.create({
@@ -34,7 +38,16 @@ export class UsersService {
   }
 
   // دریافت تمام کاربران
-  async findAll() {
+  async findAll(req: any) {
+    const { role } = req.user;
+
+    // بررسی نقش کاربر
+    if (role === 10) {
+      return {
+        statusCode: 403,
+        error: 'شما به این بخش دسترسی ندارید',
+      };
+    }
     const allUser = await this.UsersModel.find();
 
     return {
@@ -46,10 +59,12 @@ export class UsersService {
 
   // دریافت کاربر بر اساس نام کاربری
   async findByUsername(username: string) {
+    // دریافت کاربر
     const findUser = await this.UsersModel.findOne({
       username,
     });
 
+    // کاربری وجود ندارد
     if (!findUser) {
       return {
         statusCode: 400,
@@ -66,10 +81,12 @@ export class UsersService {
 
   // دریافت کاربر بر اساس نام ایمیل
   async findByEmail(email: string) {
+    // دریافت کاربر
     const findUser = await this.UsersModel.findOne({
       email,
     });
 
+    // کاربری وجود ندارد
     if (!findUser) {
       return {
         statusCode: 400,
@@ -86,10 +103,12 @@ export class UsersService {
 
   // دریافت کاربر بر اساس نام آیدی
   async findById(id: string) {
+    // دریافت یورز
     const findUser = await this.UsersModel.findOne({
       _id: id,
     });
 
+    // کاربری وجود ندارد
     if (!findUser) {
       return {
         statusCode: 400,
@@ -106,18 +125,32 @@ export class UsersService {
   F;
 
   // بروزرسانی کاربر
-  async update(username: string, updateUserDto: UpdateUserDto) {
+  async update(username: string, updateUserDto: UpdateUserDto, req?: any) {
+    const { role, username: userNameUser } = req.user;
+
+    // بررسی نقش کاربر
+    if (role === 10) {
+      // اگر کاربر به عنوان نقش ممبر باشد نباید بهتواند نقش و وضعیت یوزر خودش
+      // یا اطلاعات بقیه کاربران رو ویرایش کن
+      username = userNameUser;
+      updateUserDto.username = userNameUser;
+      updateUserDto.role = role;
+      updateUserDto.status = true;
+    }
+
     const updateUser = await this.UsersModel.updateOne(
       { username },
       { ...updateUserDto },
     );
 
-    if (!updateUser) {
+    // کاربری وجود ندارد
+    if (!updateUser.modifiedCount || updateUser.acknowledged) {
       return {
         statusCode: 400,
         error: 'کاربری با این نام کاربری وجود ندارد',
       };
     }
+
     return {
       statusCode: 201,
       message: 'کاربر با موفقیت بروزرسانی شد.',
@@ -125,8 +158,26 @@ export class UsersService {
   }
 
   // حذف کاربر
-  async remove(username: string) {
-    await this.UsersModel.deleteOne({ username });
+  async remove(username: string, req?: any) {
+    const { role } = req.user;
+
+    // بررسی نقش کاربر
+    if (role === 10) {
+      return {
+        statusCode: 403,
+        error: 'شما به این بخش دسترسی ندارید',
+      };
+    }
+    const deleteUser = await this.UsersModel.deleteOne({ username });
+
+    // کاربری وجود ندارد
+    if (!deleteUser.deletedCount) {
+      return {
+        statusCode: 400,
+        error: 'کاربری با این نام کاربری وجود ندارد',
+      };
+    }
+
     return {
       statusCode: 200,
       message: 'کاربر با موفقیت حذف شد.',
